DUCKM-36 fix: use std::fabs instead of std::fabsl
- std::fabsl is not portable in std namespace; std::fabs has long double overload

+++ storage/duckdb/duckdb_snapshot_loader.cc
@@ -655,7 +655,7 @@ bool ValidateSums(MYSQL *mysql, DuckDBAdapter &adapter,
                 << def.name << "." << col.name << "\n";
       return false;
     }
-    const long double diff = std::fabsl(mysql_sum - duckdb_sum);
+    const long double diff = std::fabs(mysql_sum - duckdb_sum);
     const long double tol = 1e-6L;
     if (diff > tol) {
       std::cerr << "Validation mismatch for " << schema << "." << def.name
